% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fx_plot_themes.R
\name{fx_plot_themes}
\alias{fx_plot_themes}
\title{Title KP's ggplot themes}
\usage{
fx_plot_themes()
}
\value{
ggplot theme for light and dark plots
}
\description{
Title KP's ggplot themes
}
\examples{
kp_themes <- fx_plot_themes()
kp_themes$light_theme
kp_themes$dark_theme
}
