% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fx_top_p.R
\name{fx_top_p}
\alias{fx_top_p}
\title{Title}
\usage{
fx_top_p(data, p, wt)
}
\arguments{
\item{data}{A data.frame}

\item{p}{The proportion to sample}

\item{wt}{The unquoted column to use for ranking}
}
\value{
A data.frame
}
\description{
Title
}
\examples{
fx_top_p(data = mtcars, p = 0.25, wt = "mpg")
}
