% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fx_plot_themes.R
\name{theme_kp}
\alias{theme_kp}
\title{Title KP's ggplot themes}
\usage{
theme_kp(style = "light")
}
\arguments{
\item{style}{Chosen style for ggplot output. Valid options are 'light' or 'dark'}
}
\value{
ggplot theme for light and dark plots
}
\description{
Title KP's ggplot themes
}
\examples{
theme_kp("light")
theme_kp("dark")
}
