% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fx_describe.R
\name{fx_describe}
\alias{fx_describe}
\title{Title}
\usage{
fx_describe(data, output_format = "character")
}
\arguments{
\item{data}{A data.frame}

\item{output_format}{Type of values the output table should have.  Default is "character" for proper display. Use "numeric" for unrounded values.}
}
\value{
character vector of names
}
\description{
Title
}
\examples{
fx_describe(mtcars)
}
