% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fx_describe.R
\name{fx_describe}
\alias{fx_describe}
\title{Title}
\usage{
fx_describe(data)
}
\arguments{
\item{data}{A data.frame}
}
\value{
character vector of names
}
\description{
Title
}
\examples{
fx_describe(mtcars)
}
